开发匹配弹框及相关接口逻辑

+++ lib_base/src/main/java/com/alan/mvvm/base/http/responsebean/GoodBean.kt
@@ -9,7 +9,7 @@ data class GoodBean(
     val goodsId: String? = null,
     val name: String? = null,
     val type: String? = null,
-    val amount: Int = 0,
+    val amount: Float = 0f,
     val point: Int = 0,
     val detail: String? = null,
     val date: String? = null,
